# docs/swagger.yaml
basePath: /
definitions:
...
  db.CardsLabel:
    properties:
      card_id:
        example: 32
        type: integer
      id:
        example: 1
        type: integer
      label_id:
        example: 64
        type: integer
    type: object
...
host: localhost:3000
info:
  contact:
    email: igorek.denisenko.03@mail.ru
    name: OLD_GREEN_FROG
  description: Service for managing projects
  title: PrushkaAPI
  version: "1.0"
paths:
  /attachment/:
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /attachment/{id}/:
    delete:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /cards/:
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /cards/{id}/:
    delete:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /cards_label/:
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /cards_label/{id}/:
    delete:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /column/:
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /column/{id}/:
    delete:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /desk/:
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /desk/{id}/:
    delete:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /label/:
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /label/{id}/:
    delete:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
  /privilege/:
    get:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
    post:
      description: CRUD for models
      parameters:
      - description: Get all models
        in: query
        name: all
        type: boolean
      - description: Get n models
        in: query
        name: limit
        type: integer
      - description: Get model with id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CardsLabel'
            type: array
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      summary: Work with all existing models
...
swagger: "2.0"